docs: update deployment documentation and fix schema for GitHub Actions
- Update service URL to https://fincore-npe-api-994490239798.europe-west2.run.app
- Fix Permission_Identifier column type from INT to BIGINT in DDL files
- Update GitHub Actions to use Secret Manager for DB_PASSWORD
- Change --cpu-throttling to --cpu-boost for better cold start
- Update GitHub secrets documentation (SECRET_NAME instead of DB_PASSWORD)
- Update version to 1.0.1 and last updated date

+++ cloud-sql-schema.sql
@@ -14,7 +14,7 @@ DROP TABLE IF EXISTS Permissions;
 
 -- Create Permissions table
 CREATE TABLE Permissions (
-    Permission_Identifier INT PRIMARY KEY AUTO_INCREMENT,
+    Permission_Identifier BIGINT PRIMARY KEY AUTO_INCREMENT,
     Permission_Name VARCHAR(100) UNIQUE NOT NULL,
     Description TEXT,
     Resource VARCHAR(50),
